Drop support for legacy format of ProposalSettings
