+++ src/lib/y2storage/proposal/base.rb
@@ -51,9 +51,6 @@ module Y2Storage
       # If the proposal was generated by any other means, this attribute should
       # be nil.
       #
-      # @note This is expected to be nil in any case if the proposal settings use
-      #   the legacy format.
-      #
       # @return [Proposal::SettingsAdjustment, nil]
       attr_accessor :auto_settings_adjustment
