--- src/lib/y2storage/proposal/devices_planner_strategies.rb
@@ -1,30 +1,0 @@
-# Copyright (c) [2017] SUSE LLC
-#
-# All Rights Reserved.
-#
-# This program is free software; you can redistribute it and/or modify it
-# under the terms of version 2 of the GNU General Public License as published
-# by the Free Software Foundation.
-#
-# This program is distributed in the hope that it will be useful, but WITHOUT
-# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
-# more details.
-#
-# You should have received a copy of the GNU General Public License along
-# with this program; if not, contact SUSE LLC.
-#
-# To contact SUSE LLC about this file by physical or electronic mail, you may
-# find current contact information at www.suse.com.
-
-module Y2Storage
-  module Proposal
-    # Namespace to group the different strategies to plan devices. The planner
-    # strategy depends on the format of the proposal settings.
-    module DevicesPlannerStrategies
-    end
-  end
-end
-
-require "y2storage/proposal/devices_planner_strategies/legacy"
-require "y2storage/proposal/devices_planner_strategies/ng"
