+++ test/y2storage/dialogs/guided_setup/select_filesystem_test.rb
@@ -21,14 +21,14 @@
 require_relative "../../../spec_helper.rb"
 require_relative "#{TEST_PATH}/support/guided_setup_context"
 
-describe Y2Storage::Dialogs::GuidedSetup::SelectFilesystem::Ng do
+describe Y2Storage::Dialogs::GuidedSetup::SelectFilesystem do
   include_context "guided setup requirements"
 
   subject(:dialog) { described_class.new(guided_setup) }
 
   before { allow(settings).to receive(:volumes).and_return volumes }
 
-  let(:widget_class) { Y2Storage::Dialogs::GuidedSetup::SelectFilesystem::VolumeWidget }
+  let(:widget_class) { Y2Storage::Dialogs::GuidedSetup::Widgets::Volume }
 
   describe "#run" do
     let(:volumes) do
